separate out the Hypermodel from AutoModel
